refactor: reorganize cli folder structure and move core modules into /core subfolder
