+++ src/cli/core/generate-path-structure.ts
