+++ src/cli/core/path-utils.test.ts
