+++ src/cli/core/route-scanner.test.ts
