+++ src/cli/core/scan-utils.test.ts
